sample_agents/requirements.txt
google-adk==1.1.1
deprecated
